1) Adding JsonCompare in comparision.ts 2) Formating document 3) Removing jons_utils.ts
